+++ changes.md
@@ -1,6 +1,7 @@
 # Change Log
 
-## v0.1.35 
+## v0.1.35 Loading PCT14 Import
+* Loading exported data from PCT14 introduced.
 * eo_man can now take EEP from TeachIn telegram and pre-configure device.
 
 ## v0.1.34 Added more configuration templates for devices